test(e2e): split hermes slack assertions into helper
